Fix bug where loop() is run for ALL sensors after an interrupt
`_last_interrupt_pin` should be reset AFTER the for loop has finished
running over all the sensors, not DURING the for loop. I have a node
with 1x SensorDoor and 1x SensorDS18b20 (with 4 children) and before
this fix, every time the SensorDoor changed state the node would also
measure and report all the DS18B20 before going back to sleep. After
this fix, the node reports SensorDoor only, then goes straight back to
sleep.

However this now raises the issue that when the node goes to sleep, it
goes to sleep for the full sleep interval, i.e. if the sleep interval is
10 minutes, and you keep triggering the SensorDoor every 5 minutes, the
node will never wake up to measure and repor the other sensors. Will
address that issue in a separate commit.